requirements.txt
numpy~=1.19.5
keras~=2.6.0
matplotlib~=3.4.3
tensorflow~=2.6.0
PyQt5~=5.15.4
Pillow~=8.3.1
scipy~=1.7.1